refactor: move otel exporter to exporter package
