+++ makefile.manifest.inc
@@ -6,7 +6,11 @@
 # Teryl Taylor <terylt@ibm.com>
 #
 
+<<<<<<< HEAD
 SYSFLOW_VERSION?=0.6.3
+=======
+SYSFLOW_VERSION?=0.7.0-rc1
+>>>>>>> 1400413 (refactor: move otel exporter to exporter package)
 SYSFLOW_BUILD_NUMBER?=1
 SYSFLOW_JSON_SCHEMA_VERSION=5
 SYSFLOW_ECS_VERSION=8.2
